[flutter_tools] remove most mocks from Fuchsia device tests
